fix account decode err when replay usage from old ckp (#13868)
1. correct the account type in the storage usage bat to uint64.
2. adding more tests to test the behavior of the usage cache.
3. one way to provide support is to transfer the usage to the right account.

Approved by: @qingxinhome, @XuPeng-SH, @sukki37

+++ pkg/vm/engine/tae/logtail/utils.go
@@ -712,7 +712,7 @@ type BaseCollector struct {
 		Deletes        []interface{}
 		SegInserts     []*catalog.ObjectEntry
 		SegDeletes     []*catalog.ObjectEntry
-		ReservedAccIds map[uint32]struct{}
+		ReservedAccIds map[uint64]struct{}
 	}
 
 	UsageMemo *TNUsageMemo
@@ -773,7 +773,7 @@ func NewGlobalCollector(end types.TS, versionInterval time.Duration) *GlobalColl
 	collector.ObjectFn = collector.VisitObj
 	collector.BlockFn = collector.VisitBlk
 
-	collector.Usage.ReservedAccIds = make(map[uint32]struct{})
+	collector.Usage.ReservedAccIds = make(map[uint64]struct{})
 
 	return collector
 }
@@ -2698,7 +2698,7 @@ func (collector *GlobalCollector) VisitDB(entry *catalog.DBEntry) error {
 	}
 
 	currAccId := uint32(entry.GetTenantID())
-	collector.Usage.ReservedAccIds[currAccId] = struct{}{}
+	collector.Usage.ReservedAccIds[uint64(currAccId)] = struct{}{}
 
 	return collector.BaseCollector.VisitDB(entry)
 }
